+++ README.md
@@ -17,15 +17,8 @@ SwiftOpenAI, you will find that we initialize the aiproxy service slightly
 differently here. We no longer accept a `deviceCheckBypass` as an argument to
 the initializer of the service. It was too easy to accidentally leak the constant.
 
-Instead, you add the device check bypass as an environment variable. Here are the
-steps:
-
-1. Type `cmd shift ,` to open up the "Edit Schemes" menu.
-2. Select `Run` in the sidebar
-3. Select `Arguments` from the top nav
-4. Add to the "Environment Variables" section (not the "Arguments Passed on
-   Launch" section) an env variable with name `AIPROXY_DEVICE_CHECK_BYPASS` and
-   value that we provided you in the AIProxy dashboard
+Instead, you add the device check bypass as an environment variable. Please follow
+the steps in the next section for adding an environment variable to your project.
 
 
 ## Adding this package as a dependency to your Xcode project
@@ -33,11 +26,20 @@ steps:
 1. Open your Xcode project
 2. Select `File > Add Package Dependencies`
 3. Punch `github.com/lzell/aiproxyswift` into the package URL bar
+4. Add the device check bypass token to your Xcode project as an environment
+   variable (you get the bypass token when you configure a project on aiproxy.pro)
+    - Type `cmd shift ,` to open up the "Edit Schemes" menu.
+    - Select `Run` in the sidebar
+    - Select `Arguments` from the top nav
+    - Add to the "Environment Variables" section (not the "Arguments Passed on
+   Launch" section) an env variable with name `AIPROXY_DEVICE_CHECK_BYPASS` and
+   value that we provided you in the AIProxy dashboard
+
 
 
 ## Example usage
 
-### Get a chat completion from a text prompt:
+### Get a chat completion from openai:
 
 
     import AIProxy
@@ -58,11 +60,7 @@ steps:
     }
 
 
-Note that you have to be in an async context for the above to work. If you are
-not using structured concurrency, wrap the code above in a `Task`
-
-
-### Get a chat completion that includes an image:
+### Send a multi-modal chat completion request to openai:
 
 
     import AIProxy
@@ -98,7 +96,41 @@ not using structured concurrency, wrap the code above in a `Task`
     }
 
 
-Note that you have to be in an async context for the above to work. If you are
-not using structured concurrency, wrap the code above in a `Task`.
+### Specify your own `clientID` to annotate requests
 
-There are simple examples in `Tests/AIProxyTests/OpenAIServiceTests.swift` that you may take inspiration from.
+If your app already has client or user IDs that you want to annotate AIProxy requests with,
+pass a second argument to the provider's service initializer. For example:
+
+    let openAIService = AIProxy.openAIService(
+        partialKey: "<the-partial-key-from-the-dashboard>",
+        clientID: "<your-id>"
+    )
+
+Requests that are made using `openAIService` will be annotated on the AIProxy backend, so that
+when you view top users, or the timeline of requests, your client IDs will be familiar.
+
+If you do not have existing client or user IDs, no problem! Leave the `clientID` argument
+out, and we'll generate IDs for you. See `AIProxyIdentifier.swift` if you would like to see
+ID generation specifics.
+
+
+### Troubleshooting
+
+#### Async function context
+
+If you use the snippets above and encounter the error
+
+    'async' call in a function that does not support concurrency
+
+it is because we assume you are in a structured concurrency context. If you encounter this
+error, you can use the escape hatch of wrapping your snippet in a `Task {}`.
+
+
+#### macOS network sandbox
+
+If you encounter the error
+
+    networkd_settings_read_from_file Sandbox is preventing this process from reading networkd settings file at "/Library/Preferences/com.apple.networkd.plist", please add an exception.
+
+Modify your macOS project settings by tapping on your project in the Xcode project tree, then
+select `Signing & Capabilities` and enable `Outgoing Connections (client)`
